Change name of app struct to fix preview target name collision (#16)
When trying to preview a swiftui struct that's a member of the
bottomless target, we'd see an error saying that the file was
not a member of `Bottomless` which is both the app's target
name and the name of the overall App struct. This change aims
to fix the preview conflating what seems to be a reserved name
for the target.