Order preferences update (#57)
Looks like bottomless nixed the ability to customize the alerts for the
various stages of an order's progress. This leaves us with the ability
to flip only the "gifs enabled" switch.

This PR removes all the viewmodel and view building for the individual
order progress settings, the warning about not having added a phone
number, and updates the gif alerts setting based upon having a number
linked to your account or not to. This respects email or sms-based
account settings.